requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
google-generativeai==0.3.2
python-dotenv==1.0.0